+++ operation.h
@@ -6,10 +6,10 @@
 #include <math.h>
 
 // Sequential distance computation
-double dist(float *U, float *V, int n);
+double dist(double *U, double *V, int n);
 
 // Vectorized (AVX) version assuming U and V are aligned and n is multiple of 8
-double vect_dist(float *U, float *V, int n);
+double vect_dist(double *U, double *V, int n);
 
 // Generalized vectorized version with relaxed alignment
 double vect_dist_gen(float *U, float *V, int n);
