+++ utils.h
@@ -5,7 +5,7 @@
 #include <immintrin.h>
 
 // Function to allocate aligned memory and initialize with random values
-void initialize_random(float *arr, int size);
+void initialize_random(double *arr, int size);
 
 // Function to print the first N elements of an array
 void print_array(float *arr, int size);
